made unnecessary requiring upstream configuration when using the agent
When using the agent, the upstream configuration are configured on the agent side.
If not configured the upstream settings when using the agent, automatically set the upstream URL that is "http://via-agent".

But the agent is an L4 proxy. front-proxy has to choose an L7 protocol.
If you need to connect by https by the agent, then you must configure the upstream.
(e.g. upstream: https://via-agent)